requirements.txt
pymongo
streamlit
streamlit-option-menu
pandas
matplotlib 
seaborn
joblib
python-dotenv
scikit-learn